Update on the App
1. Added the Notification and message screens.

2. Updated the jobs and events with the business name data.

3. Worked on the 'capitalize' util function to return an empty string if the data string passed intor it is null.

4. updated the business account model...

5. Hid some of the checks for checking if a business account profile has been completed., till BE adds the headline and the description in the API.

+++ lib/features/dashboard/screen/business/business_dashboard_job_details_screen.dart
@@ -1,4 +1,7 @@
+// ignore_for_file: use_build_context_synchronously
+
 import 'package:nodes/config/dependencies.dart';
+import 'package:nodes/features/dashboard/components/create_job_post.dart';
 import 'package:nodes/features/dashboard/components/job_analytics.dart';
 import 'package:nodes/features/dashboard/components/job_applicants.dart';
 import 'package:nodes/features/dashboard/components/job_details_business.dart';
@@ -32,6 +35,7 @@ class _BusinessJobDetailsScreenState extends State<BusinessJobDetailsScreen> {
   @override
   Widget build(BuildContext context) {
     dashCtrl = context.watch<DashboardController>();
+    job = dashCtrl.currentlyViewedBusinessJob;
     return Stack(
       children: [
         ListView(
@@ -56,15 +60,16 @@ class _BusinessJobDetailsScreenState extends State<BusinessJobDetailsScreen> {
                   ),
                   actionBtn(
                     icon: ImageUtils.editPencileOutlineIcon,
-                    onTap: () {
-                      //
-                    },
-                    loading: false,
+                    onTap: showCreateJobBottomSheet,
+                    loading: context.watch<DashboardController>().isUpdatingJob,
                   ),
                   actionBtn(
                     icon: ImageUtils.shareOutlineIcon,
-                    onTap: () {
-                      //
+                    onTap: () async {
+                      await shareDoc(
+                        context,
+                        url: "$nodeWebsite/${job.name}",
+                      );
                     },
                     loading: false,
                   ),
@@ -172,9 +177,52 @@ class _BusinessJobDetailsScreenState extends State<BusinessJobDetailsScreen> {
   }
 
   void deleteJob() async {
-    bool done = await dashCtrl.deleteSingleJob(context, job.id);
-    if (done && mounted) {
-      customNavigateBack(context);
+    final result = await showAlertDialog(
+      context,
+      body: subtext(
+        "Are you sure you want to delete this Job posting?",
+        fontSize: 13,
+      ),
+      title: "Delete Job",
+      cancelTitle: "No, Cancel",
+      okTitle: "Yes, Delete",
+      okColor: RED,
+      cancelColor: GRAY,
+    );
+    if (DialogAction.yes == result) {
+      // make the api request, delete account, call the logout function, to send them to the auth screen...
+      bool done = await dashCtrl.deleteSingleJob(context, job.id);
+      if (done && mounted) {
+        customNavigateBack(context);
+      }
     }
+  }
+
+  showCreateJobBottomSheet() {
+    showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      isDismissible: true,
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.vertical(
+          top: Radius.circular(30.0),
+        ),
+      ),
+      backgroundColor: WHITE,
+      elevation: 0.0,
+      builder: (context) {
+        return BottomSheetWrapper(
+          closeOnTap: true,
+          title: labelText(
+            "Edit Job Post",
+            fontSize: 18,
+            fontWeight: FontWeight.w500,
+          ),
+          child: CreateJobPost(
+            job: job,
+          ),
+        );
+      },
+    );
   }
 }
